+++ CKCQUIZZ.Server/Interfaces/ICauHoiService.cs
@@ -11,7 +11,5 @@ namespace CKCQUIZZ.Server.Interfaces
         Task<bool> DeleteAsync(int id);
         Task<PagedResult<CauHoiDto>> GetQuestionsForAssignedSubjectsAsync(string userId, QueryCauHoiDto query);
         Task<PagedResult<CauHoiDto>> GetMyCreatedQuestionsAsync(string userId, QueryCauHoiDto query);
-
-        Task<KetQuaImportViewModel> ImportFromZipAsync(IFormFile file,int maMonHoc,int maChuong,int doKho,string userId);
     }
 }
