+++ CKCQUIZZ.Server/appsettings.json
@@ -7,7 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=localhost;Initial Catalog=CKCQUIZZ_Final;Integrated Security=True;Trust Server Certificate=True"
+    "DefaultConnection": "Data Source=localhost;Initial Catalog=CKCQUIZZ_Final;Integrated Security=False;Trust Server Certificate=True;User=sa;Password=Thongle789321@"
   },
   "JWT": {
     "Issuer": "https://localhost:7254",
