+++ ckcquizz.client/src/views/admin/question/index.vue
@@ -323,11 +323,9 @@ import debounce from 'lodash/debounce';
 import { message, Modal } from 'ant-design-vue';
 import apiClient from '@/services/axiosServer';
 import { useUserStore } from '@/stores/userStore';
-const API_BASE_URL = apiClient.defaults.baseURL.replace('/api', '');
 const DynamicFormElements = defineAsyncComponent(() => import('./DynamicFormElements.vue'));
 
 const columns = [
-/*  { title: 'ID', dataIndex: 'macauhoi', key: 'macauhoi', width: 60 },*/
   { title: 'Nội dung câu hỏi', dataIndex: 'noidung', key: 'noidung' },
   { title: 'Loại', dataIndex: 'loaicauhoi', key: 'loaicauhoi', width: 130, align: 'center' },
   { title: 'Môn học', dataIndex: 'tenMonHoc', key: 'tenMonHoc', width: 180 },
@@ -353,7 +351,7 @@ const getInitialFormState = () => ({
   maChuong: null,
   doKho: 1,
   noidung: '',
-  hinhanhUrl: null, // Sử dụng tên thống nhất
+  hinhanhUrl: null, 
   fileList: [],
   trangthai: true,
   dapAn: [{ noidung: '' }],
@@ -362,7 +360,6 @@ const getInitialFormState = () => ({
   dapAnTuLuan: '',
 });
 
-// State cho Modal Thêm
 const isAddModalVisible = ref(false);
 const addModalLoading = ref(false);
 const addFormRef = ref();
@@ -370,7 +367,6 @@ let addFormState = reactive(getInitialFormState());
 const modalChapters = ref([]);
 const modalChaptersLoading = ref(false);
 
-// State cho Modal Sửa
 const isEditModalVisible = ref(false);
 const editModalLoading = ref(false);
 const editFormRef = ref();
@@ -378,16 +374,14 @@ let editFormState = reactive(getInitialFormState());
 const editModalChapters = ref([]);
 const editModalChaptersLoading = ref(false);
 const isEditModalInitializing = ref(false);
-  // =======================================================
-  // STATE CHO MODAL IMPORT
-  // =======================================================
+
   const isImportModalVisible = ref(false);
   const importModalLoading = ref(false);
   const importFormRef = ref();
   const importFormState = reactive({
     maMonHoc: null,
     maChuong: null,
-    doKho: 1, // Mặc định là Dễ
+    doKho: 1, 
     fileList: [],
   });
   const importModalChapters = ref([]);
@@ -398,7 +392,6 @@ const isEditModalInitializing = ref(false);
     maChuong: [{ required: true, message: 'Vui lòng chọn chương!' }],
     fileList: [{ required: true, type: 'array', min: 1, message: 'Vui lòng tải lên một file .zip!' }]
   };
-// Rules validation
 const formRules = {
   maMonHoc: [{ required: true, message: 'Vui lòng chọn môn học!' }],
   maChuong: [{ required: true, message: 'Vui lòng chọn chương!' }],
@@ -449,7 +442,6 @@ const fetchChaptersForModal = async (subjectId, targetChaptersRef, loadingRef) =
   finally { loadingRef.value = false; }
 };
 
-// CÁC HÀM XỬ LÝ SỰ KIỆN (BẢNG, LỌC, XÓA)
 const handleTableChange = (p) => {
   pagination.current = p.current;
   pagination.pageSize = p.pageSize;
@@ -508,7 +500,6 @@ const handleAddOk = async () => {
 
 const handleAddCancel = () => { isAddModalVisible.value = false; };
 
-// CÁC HÀM XỬ LÝ MODAL SỬA
 const openEditModal = async (record) => {
   isEditModalVisible.value = true;
   editModalLoading.value = true;
@@ -528,7 +519,7 @@ const openEditModal = async (record) => {
       maChuong: data.machuong,
       loaiCauHoi: data.loaicauhoi,
       hinhanhUrl: data.hinhanhurl,
-      fileList: data.hinhanhurl ? [{ uid: '-1', name: 'image.png', status: 'done', url: `${API_BASE_URL}${data.hinhanhurl}` }] : [],
+      fileList: data.hinhanhurl ? [{ uid: '-1', name: 'image.png', status: 'done', url: data.hinhanhurl }] : [],
       dapAn: data.cauTraLois.length > 0 ? data.cauTraLois.map(ans => ({ macautl: ans.macautl, noidung: ans.noidungtl })) : getInitialFormState().dapAn,
       dapAnTuLuan: data.loaicauhoi === 'essay' && data.cauTraLois.length > 0 ? data.cauTraLois[0].noidungtl : '',
       correctAnswer: getCorrectAnswerFromApi(data.cauTraLois, data.loaicauhoi),
@@ -565,7 +556,6 @@ const handleEditOk = async () => {
 
 const handleEditCancel = () => { isEditModalVisible.value = false; };
 
-// CÁC HÀM HELPER (TÁI SỬ DỤNG)
 const createPayload = (formState) => {
   const basePayload = {
     loaiCauHoi: formState.loaiCauHoi,
@@ -641,9 +631,7 @@ const handleApiError = (error, defaultMessage) => {
 
 const formatQuestionType = (type) => ({ 'single_choice': 'Một đáp án', 'multiple_choice': 'Nhiều đáp án', 'essay': 'Tự luận' }[type] || 'N/A');
 const getQuestionTypeTagColor = (type) => ({ 'single_choice': 'blue', 'multiple_choice': 'cyan', 'essay': 'purple', 'image': 'orange', }[type] || 'default');
-  // CÁC HÀM XỬ LÝ MODAL IMPORT
   const showImportModal = () => {
-    // Reset form state mỗi khi mở modal
     Object.assign(importFormState, {
       maMonHoc: null,
       maChuong: null,
@@ -659,7 +647,6 @@ const getQuestionTypeTagColor = (type) => ({ 'single_choice': 'blue', 'multiple_
   };
 
   const downloadTemplate = () => {
-    // Đảm bảo bạn có file /public/templates/mau-import.zip trong dự án Vue
     const link = document.createElement('a');
     link.href = '/templates/import_cauhoi.zip';
     link.download = "mau_cauhoi.zip";
@@ -677,13 +664,12 @@ const getQuestionTypeTagColor = (type) => ({ 'single_choice': 'blue', 'multiple_
       formData.append('file', importFormState.fileList[0].originFileObj);
 
       const { maMonHoc, maChuong, doKho } = importFormState;
-      const url = `/CauHoi/import-from-zip?maMonHoc=${maMonHoc}&maChuong=${maChuong}&doKho=${doKho}`;
+      const url = `/Files/import-from-zip?maMonHoc=${maMonHoc}&maChuong=${maChuong}&doKho=${doKho}`;
       const response = await apiClient.post(url, formData, {
         headers: { 'Content-Type': 'multipart/form-data' },
-        timeout: 300000 // Tăng timeout lên 5 phút cho các file lớn
+        timeout: 300000 
       });
 
-      // Hiển thị thông báo thành công
       message.success(response.data.thongBao || 'Import thành công!');
 
       isImportModalVisible.value = false;
@@ -701,7 +687,6 @@ const getQuestionTypeTagColor = (type) => ({ 'single_choice': 'blue', 'multiple_
           width: '600px'
         });
       } else {
-        // Các lỗi khác
         handleApiError(error, "Import thất bại");
       }
     } finally {
@@ -710,10 +695,8 @@ const getQuestionTypeTagColor = (type) => ({ 'single_choice': 'blue', 'multiple_
   };
   watch(() => importFormState.maMonHoc, (newVal) => {
     importFormState.maChuong = null;
-    // Tái sử dụng hàm fetchChaptersForModal của bạn
     fetchChaptersForModal(newVal, importModalChapters, importModalChaptersLoading);
   });
-// WATCHERS (THEO DÕI SỰ THAY ĐỔI)
 watch(() => addFormState.maMonHoc, (newVal) => {
   addFormState.maChuong = null;
   fetchChaptersForModal(newVal, modalChapters, modalChaptersLoading);
@@ -733,7 +716,6 @@ watch(() => addFormState.hasImage, (newValue) => {
     addFormState.hinhanhUrl = '';
   }
 });
-// LIFECYCLE HOOKS
 onMounted(async () => {
   await userStore.fetchUserPermissions();
   fetchData();
